move cluster commands from CLI into cluster extension
