+++ src/cluster/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "fluvio-cluster"
-version = "0.3.0"
+version = "0.3.1"
 edition = "2018"
 license = "Apache-2.0"
 authors = ["Fluvio Contributors <team@fluvio.io>"]
@@ -16,8 +16,7 @@ path = "src/lib.rs"
 default = ["native_tls"]
 rust_tls = ["fluvio/rust_tls","k8-client/rust_tls"]
 native_tls = ["fluvio/native_tls","k8-client/native_tls"]
-k8 = []
-local = []
+cmd_extension = ["structopt","fluvio-extension-common","prettytable-rs"]
 
 [dependencies]
 colored = "2.0.0"
@@ -29,7 +28,8 @@ serde = "1.0.115"
 serde_json = "1.0.57"
 thiserror = "1.0.20"
 async-trait = "0.1.21"
-
+structopt = { version = "0.3.16", default-features = false, optional = true }
+prettytable-rs = { version = "0.8.0", optional = true }
 
 # Fluvio dependencies
 fluvio-helm = { version = "0.1.0", git = "https://github.com/infinyon/fluvio-helm", branch = "main"}
@@ -41,6 +41,7 @@ k8-config = { version = "1.3.0", features = ["context"] }
 k8-client = { version = "3.0.0", default-features = false }
 k8-obj-core = "1.1.0"
 k8-obj-metadata = "1.0.0"
+fluvio-extension-common = { version = "0.1.0", path = "../extension-common", optional = true }
 
 [dev-dependencies]
 fluvio-future = { version = "0.1.0", features =["task"]}
