+++ src/cluster/src/extension/cluster/releases.rs
@@ -1,6 +1,6 @@
 use structopt::StructOpt;
-use fluvio_cluster::ClusterInstaller;
-use crate::Result;
+use crate::ClusterInstaller;
+use super::Result;
 
 #[derive(Debug, StructOpt)]
 pub enum ReleasesCmd {
