+++ src/cluster/src/extension/group/delete.rs
