+++ src/cluster/src/extension/spu/display.rs
