+++ src/cluster/src/extension/spu/list.rs
