+++ src/cluster/src/extension/spu/register.rs
