+++ src/cluster/src/extension/spu/unregister.rs
